requirements.txt
Django==6.0.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
dj-database-url==3.0.1
